test: consolidate device tracker deprecation checks
